# music_player_backend/interfaces/backend_live_openapi.json
{"openapi":"3.1.0","info":{"title":"Music Player Backend API","description":"Backend for a personal music library.\n\nAuthentication:\n- Obtain a token via POST /auth/login (or /auth/register)\n- Send it on subsequent requests: Authorization: Bearer <token>\n\nStreaming:\n- GET /songs/{song_id}/stream requires Authorization header and only serves owned songs.","version":"1.0.0"},"paths":{"/auth/register":{"post":{"tags":["Auth"],"summary":"Register a new user","description":"Creates a new user and returns a JWT token.","operationId":"register_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","description":"Validates credentials and returns a JWT token.","operationId":"login_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/songs":{"get":{"tags":["Songs"],"summary":"List current user's songs","description":"Returns songs for the authenticated user, newest first.","operationId":"list_songs","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SongResponse"},"type":"array","title":"Response List Songs"}}}}},"security":[{"HTTPBearer":[]}]}},"/songs/upload":{"post":{"tags":["Songs"],"summary":"Upload an mp3","description":"Uploads an mp3 file for the authenticated user. Stores file on disk and metadata in DB.","operationId":"upload_song","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_song"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SongUploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/songs/{song_id}/stream":{"get":{"tags":["Songs"],"summary":"Stream a song (secure)","description":"Streams the mp3 file for the authenticated user if they own it.","operationId":"stream_song","security":[{"HTTPBearer":[]}],"parameters":[{"name":"song_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Song Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}},"audio/mpeg":{}}},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Auth"],"summary":"Health check","description":"Simple health check endpoint.","operationId":"health_check__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AuthLoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"User email address."},"password":{"type":"string","title":"Password","description":"User password."}},"type":"object","required":["email","password"],"title":"AuthLoginRequest"},"AuthRegisterRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"User email address (unique)."},"password":{"type":"string","minLength":6,"title":"Password","description":"User password (min 6 chars)."}},"type":"object","required":["email","password"],"title":"AuthRegisterRequest"},"AuthTokenResponse":{"properties":{"token":{"type":"string","title":"Token","description":"JWT access token."},"token_type":{"type":"string","title":"Token Type","description":"Token type for Authorization header.","default":"bearer"}},"type":"object","required":["token"],"title":"AuthTokenResponse"},"Body_upload_song":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"MP3 file upload (multipart/form-data)"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title","description":"Optional title. Defaults to original filename stem."},"artist":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Artist","description":"Optional artist. Defaults to 'Unknown Artist'."}},"type":"object","required":["file"],"title":"Body_upload_song"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SongResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"Song UUID."},"title":{"type":"string","title":"Title","description":"Song title."},"artist":{"type":"string","title":"Artist","description":"Song artist."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Creation timestamp."},"size_bytes":{"type":"integer","title":"Size Bytes","description":"File size in bytes."},"content_type":{"type":"string","title":"Content Type","description":"Content type stored at upload time."}},"type":"object","required":["id","title","artist","created_at","size_bytes","content_type"],"title":"SongResponse"},"SongUploadResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"Created song UUID."},"title":{"type":"string","title":"Title","description":"Song title."},"artist":{"type":"string","title":"Artist","description":"Song artist."},"filename":{"type":"string","title":"Filename","description":"Stored filename (server-side)."},"content_type":{"type":"string","title":"Content Type","description":"Uploaded content type."},"size_bytes":{"type":"integer","title":"Size Bytes","description":"Size in bytes."},"duration_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Seconds","description":"Duration in seconds if known."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Creation timestamp."}},"type":"object","required":["id","title","artist","filename","content_type","size_bytes","created_at"],"title":"SongUploadResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}},"tags":[{"name":"Auth","description":"User registration and login (JWT)."},{"name":"Songs","description":"Upload, list, and stream user-owned mp3 files."}]}